Move Supervisor's log writer into the internal log package (2/2)
This allows the writer to be used in other locations, as well.

This commit actually changes the files.
